build: 升级2025.01

+++ taotao-cloud-job/taotao-cloud-job-server/src/main/java/com/taotao/cloud/job/server/jobserver/core/schedule/DispatchService.java
@@ -1,21 +1,49 @@
+/*
+ * Copyright (c) 2020-2030, Shuigedeng (981376577@qq.com & https://blog.taotaocloud.top/).
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package com.taotao.cloud.job.server.jobserver.core.schedule;
 
-
+import static com.taotao.cloud.job.common.enums.InstanceStatus.CANCELED;
+import static com.taotao.cloud.job.common.enums.InstanceStatus.RUNNING;
+import static com.taotao.cloud.job.common.enums.InstanceStatus.WAITING_DISPATCH;
+import static com.taotao.cloud.job.common.enums.InstanceStatus.WAITING_WORKER_RECEIVE;
 
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
+import com.taotao.cloud.job.common.SystemInstanceResult;
+import com.taotao.cloud.job.common.enums.TimeExpressionType;
+import com.taotao.cloud.job.server.jobserver.common.Holder;
+import com.taotao.cloud.job.server.jobserver.common.grpc.ServerScheduleJobRpcClient;
+import com.taotao.cloud.job.server.jobserver.common.module.WorkerInfo;
+import com.taotao.cloud.job.server.jobserver.extension.lock.LockService;
+import com.taotao.cloud.job.server.jobserver.persistence.domain.InstanceInfo;
+import com.taotao.cloud.job.server.jobserver.persistence.domain.JobInfo;
 import com.taotao.cloud.job.server.jobserver.persistence.mapper.InstanceInfoMapper;
+import com.taotao.cloud.job.server.jobserver.remote.worker.WorkerClusterQueryService;
+import com.taotao.cloud.job.server.jobserver.remote.worker.selector.TaskTrackerSelectorService;
+import com.taotao.cloud.remote.protos.ScheduleCausa;
+import java.util.ArrayList;
+import java.util.Date;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-
-
 import org.springframework.stereotype.Service;
 import org.springframework.util.CollectionUtils;
-
-
-import java.util.*;
-import java.util.stream.Collectors;
-
-
 
 /**
  * 派送服务（将任务从Server派发到Worker）
@@ -30,13 +58,12 @@ import java.util.stream.Collectors;
 public class DispatchService {
 
     private final WorkerClusterQueryService workerClusterQueryService;
-//    private final InstanceManager instanceManager;
-//    private final InstanceMetadataService instanceMetadataService;
+    //    private final InstanceManager instanceManager;
+    //    private final InstanceMetadataService instanceMetadataService;
     private final InstanceInfoMapper instanceInfoMapper;
     private final TaskTrackerSelectorService taskTrackerSelectorService;
     private final ServerScheduleJobRpcClient serverScheduleJobRpcClient;
     private final LockService lockService;
-
 
     /**
      * 将任务从Server派发到Worker（TaskTracker）
@@ -55,11 +82,16 @@ public class DispatchService {
      * @param overloadOptional     超载信息，可选
      * @param appName2JobNum
      */
-    public void dispatch(JobInfo jobInfo, Long instanceId, Optional<InstanceInfo> instanceInfoOptional, Optional<Holder<Boolean>> overloadOptional, Map<String, Integer> appName2JobNum) {
+    public void dispatch(
+            JobInfo jobInfo,
+            Long instanceId,
+            Optional<InstanceInfo> instanceInfoOptional,
+            Optional<Holder<Boolean>> overloadOptional,
+            Map<String, Integer> appName2JobNum) {
         // 允许从外部传入实例信息，减少 io 次数
         // 检查当前任务是否被取消
-        InstanceInfo instanceInfo = instanceInfoMapper.selectOne(new QueryWrapper<InstanceInfo>()
-                .lambda().eq(InstanceInfo::getInstanceId, instanceId));
+        InstanceInfo instanceInfo = instanceInfoMapper.selectOne(
+                new QueryWrapper<InstanceInfo>().lambda().eq(InstanceInfo::getInstanceId, instanceId));
         Long jobId = instanceInfo.getJobId();
         if (CANCELED.getV() == instanceInfo.getStatus()) {
             log.info("[Dispatcher-{}|{}] cancel dispatch due to instance has been canceled", jobId, instanceId);
@@ -72,14 +104,21 @@ public class DispatchService {
         }
         // 任务信息已经被删除
         if (jobInfo.getId() == null) {
-            log.warn("[Dispatcher-{}|{}] cancel dispatch due to job(id={}) has been deleted!", jobId, instanceId, jobId);
-//            instanceManager.processFinishedInstance(instanceId, instanceInfo.getWfInstanceId(), FAILED, "can't find job by id " + jobId);
+            log.warn(
+                    "[Dispatcher-{}|{}] cancel dispatch due to job(id={}) has been deleted!", jobId, instanceId, jobId);
+            //            instanceManager.processFinishedInstance(instanceId, instanceInfo.getWfInstanceId(), FAILED,
+            // "can't find job by id " + jobId);
             return;
         }
 
         Date now = new Date();
         String dbInstanceParams = instanceInfo.getInstanceParams() == null ? "" : instanceInfo.getInstanceParams();
-        log.info("[Dispatcher-{}|{}] start to dispatch job: {};instancePrams: {}.", jobId, instanceId, jobInfo, dbInstanceParams);
+        log.info(
+                "[Dispatcher-{}|{}] start to dispatch job: {};instancePrams: {}.",
+                jobId,
+                instanceId,
+                jobInfo,
+                dbInstanceParams);
 
         // 查询当前运行的实例数
         long current = System.currentTimeMillis();
@@ -93,15 +132,25 @@ public class DispatchService {
         if (maxInstanceNum > 0) {
             // 不统计 WAITING_DISPATCH 的状态：使用 OpenAPI 触发的延迟任务不应该统计进去（比如 delay 是 1 天）
             // 由于不统计 WAITING_DISPATCH，所以这个 runningInstanceCount 不包含本任务自身
-            Integer runningInstanceCount = instanceInfoMapper.selectCount(new QueryWrapper<InstanceInfo>()
-                    .lambda().eq(InstanceInfo::getStatus, WAITING_WORKER_RECEIVE.getV())
-                    .or().eq(InstanceInfo::getStatus, RUNNING.getV()));
+            Long runningInstanceCount = instanceInfoMapper.selectCount(new QueryWrapper<InstanceInfo>()
+                    .lambda()
+                    .eq(InstanceInfo::getStatus, WAITING_WORKER_RECEIVE.getV())
+                    .or()
+                    .eq(InstanceInfo::getStatus, RUNNING.getV()));
             // 超出最大同时运行限制，不执行调度
             if (runningInstanceCount >= maxInstanceNum) {
-                String result = String.format(SystemInstanceResult.TOO_MANY_INSTANCES, runningInstanceCount, maxInstanceNum);
-                log.warn("[Dispatcher-{}|{}] cancel dispatch job due to too much instance is running ({} > {}).", jobId, instanceId, runningInstanceCount, maxInstanceNum);
-//                instanceInfoRepository.update4TriggerFailed(instanceId, FAILED.getV(), current, current, RemoteConstant.EMPTY_ADDRESS, result, now);
-//                instanceManager.processFinishedInstance(instanceId, instanceInfo.getWfInstanceId(), FAILED, result);
+                String result =
+                        String.format(SystemInstanceResult.TOO_MANY_INSTANCES, runningInstanceCount, maxInstanceNum);
+                log.warn(
+                        "[Dispatcher-{}|{}] cancel dispatch job due to too much instance is running ({} > {}).",
+                        jobId,
+                        instanceId,
+                        runningInstanceCount,
+                        maxInstanceNum);
+                //                instanceInfoRepository.update4TriggerFailed(instanceId, FAILED.getV(), current,
+                // current, RemoteConstant.EMPTY_ADDRESS, result, now);
+                //                instanceManager.processFinishedInstance(instanceId, instanceInfo.getWfInstanceId(),
+                // FAILED, result);
                 return;
             }
         }
@@ -110,8 +159,10 @@ public class DispatchService {
 
         if (CollectionUtils.isEmpty(suitableWorkers)) {
             log.warn("[Dispatcher-{}|{}] cancel dispatch job due to no worker available", jobId, instanceId);
-//            instanceInfoRepository.update4TriggerFailed(instanceId, FAILED.getV(), current, current, RemoteConstant.EMPTY_ADDRESS, SystemInstanceResult.NO_WORKER_AVAILABLE, now);
-//            instanceManager.processFinishedInstance(instanceId, instanceInfo.getWfInstanceId(), FAILED, SystemInstanceResult.NO_WORKER_AVAILABLE);
+            //            instanceInfoRepository.update4TriggerFailed(instanceId, FAILED.getV(), current, current,
+            // RemoteConstant.EMPTY_ADDRESS, SystemInstanceResult.NO_WORKER_AVAILABLE, now);
+            //            instanceManager.processFinishedInstance(instanceId, instanceInfo.getWfInstanceId(), FAILED,
+            // SystemInstanceResult.NO_WORKER_AVAILABLE);
             return;
         }
         // 判断是否超载，在所有可用 worker 超载的情况下直接跳过当前任务
@@ -123,7 +174,8 @@ public class DispatchService {
             return;
         }
         // todo 这里可能涉及到分片，所以有多个ip，实际只传一个
-        List<String> workerIpList = suitableWorkers.stream().map(WorkerInfo::getAddress).collect(Collectors.toList());
+        List<String> workerIpList =
+                suitableWorkers.stream().map(WorkerInfo::getAddress).collect(Collectors.toList());
         // 构造任务调度请求
         // 发送请求（不可靠，需要一个后台线程定期轮询状态,后续可能是instance的线程定时检查）
         WorkerInfo taskTracker = taskTrackerSelectorService.select(jobInfo, instanceInfo, suitableWorkers);
@@ -131,18 +183,25 @@ public class DispatchService {
 
         sendScheduleInfo(jobInfo, instanceInfo, taskTrackerAddress);
 
-        log.info("[Dispatcher-{}|{}] send schedule request to TaskTracker[address:{}] successfully.", jobId, instanceId, taskTrackerAddress);
+        log.info(
+                "[Dispatcher-{}|{}] send schedule request to TaskTracker[address:{}] successfully.",
+                jobId,
+                instanceId,
+                taskTrackerAddress);
 
         // 修改状态
-        InstanceInfo build = InstanceInfo.builder().id(instanceId).status(WAITING_WORKER_RECEIVE.getV()).taskTrackerAddress(taskTrackerAddress).build();
+        InstanceInfo build = InstanceInfo.builder()
+                .id(instanceId)
+                .status(WAITING_WORKER_RECEIVE.getV())
+                .taskTrackerAddress(taskTrackerAddress)
+                .build();
         instanceInfoMapper.updateById(build);
 
         // 若是该appName下的最后一个任务，释放appName锁
         appName2JobNum.put(jobInfo.getAppName(), appName2JobNum.get(jobInfo.getAppName()) - 1);
-        if(appName2JobNum.get(jobInfo.getAppName()) == 0){
+        if (appName2JobNum.get(jobInfo.getAppName()) == 0) {
             lockService.unlock(jobInfo.getAppName());
         }
-
     }
 
     private void sendScheduleInfo(JobInfo jobInfo, InstanceInfo instanceInfo, String taskTrackerAddress) {
@@ -154,18 +213,18 @@ public class DispatchService {
                 .setProcessorInfo(jobInfo.getProcessorInfo())
                 .setWorkerAddress(taskTrackerAddress)
                 .setTimeExpression(jobInfo.getTimeExpression())
-                .setTimeExpressionType(TimeExpressionType.of(jobInfo.getTimeExpressionType()).name())
+                .setTimeExpressionType(
+                        TimeExpressionType.of(jobInfo.getTimeExpressionType()).name())
                 .build();
 
         serverScheduleJobRpcClient.call(build);
-
     }
 
     private List<WorkerInfo> filterOverloadWorker(List<WorkerInfo> suitableWorkers) {
 
         List<WorkerInfo> res = new ArrayList<>(suitableWorkers.size());
         for (WorkerInfo suitableWorker : suitableWorkers) {
-            if (suitableWorker.overload()){
+            if (suitableWorker.overload()) {
                 continue;
             }
             res.add(suitableWorker);
